backend/requirements.txt
Django==3.2
djangorestframework>=3.12.2,<3.13.0
psycopg2>=2.8.6,<2.9.0
pillow==8.2.0
django-cors-headers==3.7.0
django-colorfield==0.4.1